Restructure coreg tests
