+++ tests/test_coreg/test_filters.py
@@ -1,0 +1,1 @@
+"""Functions to test the coregistration filters."""
